Revert "feat: post process AZAddSecret edges from roles instead of principals…"
This reverts commit e1b3b8a9634693aa83db10157e156e3781ee08eb.
