--- cmd/api/src/test/integration/harnesses/AZAddSecretHarness.json
@@ -1,186 +1,0 @@
-{
-    "style": {
-        "font-family": "sans-serif",
-        "background-color": "#ffffff",
-        "background-image": "",
-        "background-size": "100%",
-        "node-color": "#ffffff",
-        "border-width": 4,
-        "border-color": "#68bc00",
-        "radius": 50,
-        "node-padding": 5,
-        "node-margin": 2,
-        "outside-position": "auto",
-        "node-icon-image": "",
-        "node-background-image": "",
-        "icon-position": "inside",
-        "icon-size": 64,
-        "caption-position": "inside",
-        "caption-max-width": 200,
-        "caption-color": "#000000",
-        "caption-font-size": 50,
-        "caption-font-weight": "normal",
-        "label-position": "inside",
-        "label-display": "pill",
-        "label-color": "#000000",
-        "label-background-color": "#ffffff",
-        "label-border-color": "#000000",
-        "label-border-width": 4,
-        "label-font-size": 40,
-        "label-padding": 5,
-        "label-margin": 4,
-        "directionality": "directed",
-        "detail-position": "inline",
-        "detail-orientation": "parallel",
-        "arrow-width": 5,
-        "arrow-color": "#aea1ff",
-        "margin-start": 5,
-        "margin-end": 5,
-        "margin-peer": 20,
-        "attachment-start": "normal",
-        "attachment-end": "normal",
-        "relationship-icon-image": "",
-        "type-color": "#000000",
-        "type-background-color": "#ffffff",
-        "type-border-color": "#000000",
-        "type-border-width": 0,
-        "type-font-size": 16,
-        "type-padding": 5,
-        "property-position": "outside",
-        "property-alignment": "colon",
-        "property-color": "#000000",
-        "property-font-size": 16,
-        "property-font-weight": "normal"
-    },
-    "nodes": [
-        {
-            "id": "n1",
-            "position": {
-                "x": 597.3793091010091,
-                "y": -206.05591982464335
-            },
-            "caption": "AZTenant",
-            "labels": [],
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n3",
-            "position": {
-                "x": 850.2593421152671,
-                "y": -372.9391447378271
-            },
-            "caption": "AppAdminRole",
-            "labels": [],
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n4",
-            "position": {
-                "x": 1187.6561016316446,
-                "y": -307.91634473234114
-            },
-            "caption": "AZApp",
-            "labels": [],
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n5",
-            "position": {
-                "x": 850.2593421152671,
-                "y": -60.05556469118005
-            },
-            "caption": "CloudAppAdminRole",
-            "labels": [],
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n8",
-            "position": {
-                "x": 1187.6561016316446,
-                "y": -104.19549491694585
-            },
-            "caption": "AZServicePrincipal",
-            "labels": [],
-            "properties": {},
-            "style": {}
-        }
-    ],
-    "relationships": [
-        {
-            "id": "n3",
-            "fromId": "n1",
-            "toId": "n4",
-            "type": "AZContains",
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n4",
-            "fromId": "n3",
-            "toId": "n4",
-            "type": "AZAddSecret",
-            "properties": {},
-            "style": {
-                "arrow-color": "#fda1ff"
-            }
-        },
-        {
-            "id": "n9",
-            "fromId": "n1",
-            "toId": "n8",
-            "type": "AZContains",
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n21",
-            "fromId": "n1",
-            "toId": "n3",
-            "type": "AZContains",
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n22",
-            "fromId": "n1",
-            "toId": "n5",
-            "type": "AZContains",
-            "properties": {},
-            "style": {}
-        },
-        {
-            "id": "n30",
-            "fromId": "n5",
-            "toId": "n4",
-            "type": "AZAddSecret",
-            "properties": {},
-            "style": {
-                "arrow-color": "#fda1ff"
-            }
-        },
-        {
-            "id": "n32",
-            "type": "AZAddSecret",
-            "style": {
-                "arrow-color": "#fda1ff"
-            },
-            "properties": {},
-            "fromId": "n3",
-            "toId": "n8"
-        },
-        {
-            "id": "n33",
-            "type": "AZAddSecret",
-            "style": {
-                "arrow-color": "#fda1ff"
-            },
-            "properties": {},
-            "fromId": "n5",
-            "toId": "n8"
-        }
-    ]
-}
